refactor: move helm-related logic into separate files
